Cuas.jl Documentation

The anti aerial package.

Table of contents

Cuas.jl Documentation

Table of contents

Introduction

Package Features

Introduction

Cuas is a anti aerial simulation package designed to compute efficiency of a weapon system against an aerial target. It offers the following scenarios: stationary vs stationary, stationary vs moving, moving vs stationary and moving vs moving. For one on one, one on many, many on one and many on many configurations.

julia> using Cuas

Package Features

Hit probability

Fire Control System

Fire Control System

acquisition(sim,target, tank, proj, aero,fixedError,varibleErrorIn,variableErrorOut,randomError,zones, shapes; optional arguments)

Is a simJulia process which simulate the firing chain process for an ABM projectile. From target acquisition to damage assessment.
Input parameters are :
* sim is the simulation environment
* target is a target object which contains target information
* tank is a tank object (WeaponSystems) 
* proj contains the projectile informaiton
* aero contains the aerodynamic coefficient
* fixedError is a dictionary which contains the fixed bias (in mils) measured in a plane perpendicular to the LOS
* varibleErrorIn is a dictionary which contains the standard deviation of measured quantities that are propagated via Monte Carlo
* variableErrorOut is a dictionary which contains variable bias (in mils) measured in a plane perpendicular to the LOS 
* randomError is a dictionary which contains random errors (in mils) measured in a plane perpendicular to the LOS
* zones is an object of ZoneList (ExternalBallistics) which contains informaition on generated fragments
* shapes is a vector which contains the shapes of the fragments

Optional arguments are:
* Δt_fireCommand is the time delay for issuing the fire command default value is 1.0
* Δ_slew is the time delay to lay the canon in a desired position. The default value is 2.0
* Δ_aim is the time delay to aim at the target. The default value is  2.0
* Δ_range is the time delay to get the target range. The default value is 0
* Δ_fireControl is the time delay to get the ballistic correction. The default value is 1.0
* Δ_fireDemand is the time delay for the fire demand. The default value is 0.0
* Δ_fireDelay is the time delay for the fire. The default value is  0.0
* Δ_shotExit is the time delay for the shot to exit. The default value is 1.0
* w is the wind (cross and range). The default value is  Wind(0.0,0.0)
* deltaR defines the detonation distance. It expressed in projectile revolution. To the number of revolutions achieved 
to hit the target, deltaR is soustracted. The default value is 5
* N is the number of Monte Carlo simulations. The defaults is 100
* p defines the number in % of fragment that hit target to be considered a hit

Function documentation

Cuas.CEP — Method

CEP(σx,σy)

Returns the circular error probable (CEP) non equal variances (σx and σy)

\[CEP \cong \sqrt{(\sigma_x^2+\sigma_y^2)\big(1-\frac{2(\sigma_x^4+\sigma_y^4)}{9(\sigma_x^2+\sigma_y^2)^2}\big)^3}\]

Cuas.CEP — Method

CEP(σ)

Returns the circular error probable (CEP) zero offset and equal variances (σ)

\[CEP = 1.1774\sigma\]

Cuas.CEP — Method

CEP(error::RectErrorB;args)

Returns the circular error probable (CEP) The mean point of impact can be different from (0,0). Optinal arguments are:

N the population of the sample size default value is 1e6

p the probability default value is 50

Cuas.MSHP — Method

MSHP(target,error,n)

Returns the multiple shot hit probability (mshp) of n rounds.

\[P(n shots) = 1-(1-P)^n\]

Cuas.SEP — Method

SEP(σ)

Returns the circular error probable (CEP) zero offset and equal variances (σ)

\[\]

Cuas.SSHP — Method

SSHP(target, error)

Returns the single shot hit probability (sshp) for a circular target
zero mean uncorrelated
Vertical and horizontal errors are assumed to be independent...

\[P(R)=\int_{r=0}^{r=R} \int_{\theta=0}^{\theta=2\pi} \frac{1}{2\pi\sigma_x\sigma_y} \text{exp}-[\frac{(rcos\theta)^2}{2\sigma_x^2}+\frac{(rsin\theta)^2}{2\sigma_y^2}]rdrd\theta\]

Cuas.SSHP — Method

SSHP(target,error)

Returns the single shot hit probability (sshp) for a rectangular target For the side of the rectangle standard deviations mean point of impact are defined in with "error" variable. Vertical and horizontal errors are assumed to be independent...

\[\int_{-a}^{a}\int_{-b}^{b}\frac{1}{2\pi\sigma_x\sigma_y}\text{exp}[\frac{(x-\mu_x)^2}{2\sigma_x^2}\frac{(y-\mu_y)^2}{2\sigma_y^2}]dxdy\]

Cuas.SSHP — Method

SSHP(target, error)

Returns the single shot hit probability (sshp) for a spherical target
zero mean uncorrelated
Vertical and horizontal errors are assumed to be independent...

\[P(R) = \frac{1}{(2\pi)^(3/2)\sigma_x^2\sigma_y^2\sigma_z^2}\int_{0}^{R}\int_{0}^{\pi}\int_{0}^{2\pi}f(r,\phi,\theta)r^2sin\theta d\theta d\phi dr\]

where

f(r,\phi,\theta) = exp-[\frac{(r^2sin^2\phi cos^2\theta)}{2\sigma_x^2}+\frac{(r^2sin^2\theta sin^2\phi)}{2\sigma_y^2}+\frac{r^2cos^2\phi}{2\sigma_z^2}]rdrd\theta

Cuas.createProjectile — Method

createProjectile(mass,Calibre;optional arguments)

Creates an object projectile with mass (kg), calibre (m). optional arguments are: position, velocity, time of flight, moment of inertia, distance to the center of gravity.

Cuas.createTourelle — Method

createTourelle(θ,ξ;optional arguments)

Creates an object turrets with elevation angle = θ and azimuth angle = ξ (in degrees). optional arguments are angles (in degrees): Φ, Ψ, χ and rwY.

Cuas.dispersion — Method

dispersion(target,tank, proj, aero,fixedError,varibleErrorIn,variableErrorOut,randomError;optional....) Computes the fixed bias, variable bias and random errors. Returns 3 quantities:

fixed bias an object of type 'Error' which contains the fixed bias in a vertical plane to the LOS

Variable bias an object of type 'Error3D' which contains the variable bias.

The error is expressed in the range direction and two other directions perpendicular to the LOS

random error object of type 'Error' which contains the random error in a vertical plane perpendicular to the LOS

Some error are obtained from interpolations on measured data others by Monte Carlo propagation. The number of Monte Carlo simulations is N. It is an optional input its default value is 100. Each error source is propagated individualy the total error is obtained by assumption of non-correlated errors. Inputs are:

target is a target object (ExternalBallistics) it contains the position (range of the target)

tank is a tank object defined using WeaponSystems package

proj is a projectile object which contains the projectile characteristics (ExternalBallistics)

aero contains the aerodynamic coefficient of the projectile

fixedError is a dictionary which contains the fixed bias in mils

varibleErrorIn is a dictionary which contains the standard deviation on measured values

variableErrorOut is a dictionary which contains variable bias in mils measured on plane vertical to the LOS

randomError is a dictionary which contains the random error in mils

Other optional arguments are:

The wind magnitude, its default value is null.

The atmosphere default are pressure 101325Pa, temperature =15°C and relative humidity = 80.

Cuas.pHit — Method

pHit(target,error)

Returns the probability of hitting the target defined by'target' if the estimated error is provided by 'error'.

\[\int_{-a}^{a}\int_{-b}^{b}\frac{1}{2\pi\sigma_x\sigma_y}\text{exp}[\frac{(x-\mu_x)^2}{2\sigma_x^2}\frac{(y-\mu_y)^2}{2\sigma_y^2}]dxdy\]

Cuas.phitAbm — Method

phitAbm(tank,t, proj, w,atmosphere, fuze, aero,zones, shapes,dfError,linError;optinal...) Computes the hit probability of an air bursting munition. Returns 3 quantities:

The probability of hitting the target defined by'target' with 'p %' number of fragments.

(p is an optional input, default value is 10).

A vector containing the detonation points coordinates

A vector containing the number of impacting fragments per detonation

The computation are done using MonteCarlo simulations. The number of Monte Carlo simulations is N. It is an optional input its default value is 100. Inputs are:

tank is a tank object defined using WeaponSystems package

t is a target object (ExternalBallistics) it contains the position (range of the target)

proj is a projectile object which contains the projectile characteristics (ExternalBallistics)

w is the wind magnitude (cross and range wind)

atmosphere contains the air characteristics

fuze defines the number of revolutions before detonating the projectile

aero contains the aerodynamic coefficient of the projectile

zones is a list of zones that defines the properties of the generated fragments

shapes defines the shapes of the fragments

dfError is a dictionary which contains the errors on input parameters that need to be propagated

linError is a dictionary which contains the errors measured that are tabulated

Cuas.sshpAbm — Method

sshpAbm(tank,t, proj, fuze, aero,zones, shapes,ϵ;optional....) Computes the hit probability of an air bursting munition. Returns 3 quantities:

The probability of hitting the target defined by'target' with 'p %' number of fragments.

(p is an optional input, default value is 10).

A vector containing the detonation points coordinates

A vector containing the number of impacting fragments per detonation

The computation are done using MonteCarlo simulations. The number of Monte Carlo simulations is N. It is an optional input its default value is 100. Inputs are:

tank is a tank object defined using WeaponSystems package

t is a target object (ExternalBallistics) it contains the position (range of the target)

proj is a projectile object which contains the projectile characteristics (ExternalBallistics)

fuze defines the number of revolutions before detonating the projectile

aero contains the aerodynamic coefficient of the projectile

zones is a list of zones that defines the properties of the generated fragments

shapes defines the shapes of the fragments

ϵ contains the total error defined using a 'SpheError' object.

Type documentation

Cuas.Target — Type

Target(θ,ϕ) defines a target

t0 : the time at which the target appears

ρ : the range to the target in m

θ : the elevation of the LOS

ϕ : the azimuth of the LOS

ρ′ : target velocity in range

θ′ : target velocity in elevation

ϕ′ : target velocity in azimuth

ρ″ : target acceleration in range

θ″ : target acceleration in elelvation

ϕ″ : target accelertion in azimuth

type : type of target

Cuas.Target

Cuas.CEP

Cuas.CEP

Cuas.CEP

Cuas.MSHP

Cuas.SEP

Cuas.SSHP

Cuas.SSHP

Cuas.SSHP

Cuas.createProjectile

Cuas.createTourelle

Cuas.dispersion

Cuas.pHit

Cuas.phitAbm

Cuas.sshpAbm