Cuas.jl Documentation

The anti aerial package.

Table of contents

Cuas.jl Documentation

Table of contents

Introduction

Package Features

Introduction

Cuas is a anti aerial simulation package designed to compute efficiency of a weapon system against an aerial target. It offers the following scenarios: stationary vs stationary, stationary vs moving, moving vs stationary and moving vs moving. For one on one, one on many, many on one and many on many configurations.

julia> using Cuas

Package Features

Hit probability

Function documentation

Cuas.CEP — Method

CEP(σx,σy)

Returns the circular error probable (CEP) non equal variances (σx and σy)

\[CEP \cong \sqrt{(\sigma_x^2+\sigma_y^2)\big(1-\frac{2(\sigma_x^4+\sigma_y^4)}{9(\sigma_x^2+\sigma_y^2)^2}\big)^3}\]

Cuas.CEP — Method

CEP(σ)

Returns the circular error probable (CEP) zero offset and equal variances (σ)

\[CEP = 1.1774\sigma\]

Cuas.CEP — Method

CEP(error::RectErrorB;args)

Returns the circular error probable (CEP) The mean point of impact can be different from (0,0). Optinal arguments are:

N the population of the sample size default value is 1e6

p the probability default value is 50

Cuas.MSHP — Method

MSHP(target,error,n)

Returns the multiple shot hit probability (mshp) of n rounds.

\[P(n shots) = 1-(1-P)^n\]

Cuas.SEP — Method

SEP(σ)

Returns the circular error probable (CEP) zero offset and equal variances (σ)

\[\]

Cuas.SSHP — Method

SSHP(target, error)

Returns the single shot hit probability (sshp) for a circular target
zero mean uncorrelated
Vertical and horizontal errors are assumed to be independent...

\[P(R)=\int_{r=0}^{r=R} \int_{\theta=0}^{\theta=2\pi} \frac{1}{2\pi\sigma_x\sigma_y} \text{exp}-[\frac{(rcos\theta)^2}{2\sigma_x^2}+\frac{(rsin\theta)^2}{2\sigma_y^2}]rdrd\theta\]

Cuas.SSHP — Method

SSHP(target,error)

Returns the single shot hit probability (sshp) for a rectangular target For the side of the rectangle standard deviations mean point of impact are defined in with "error" variable. Vertical and horizontal errors are assumed to be independent...

\[\int_{-a}^{a}\int_{-b}^{b}\frac{1}{2\pi\sigma_x\sigma_y}\text{exp}[\frac{(x-\mu_x)^2}{2\sigma_x^2}\frac{(y-\mu_y)^2}{2\sigma_y^2}]dxdy\]

Cuas.SSHP — Method

SSHP(target, error)

Returns the single shot hit probability (sshp) for a spherical target
zero mean uncorrelated
Vertical and horizontal errors are assumed to be independent...

\[P(R) = \frac{1}{(2\pi)^(3/2)\sigma_x^2\sigma_y^2\sigma_z^2}\int_{0}^{R}\int_{0}^{\pi}\int_{0}^{2\pi}f(r,\phi,\theta)r^2sin\theta d\theta d\phi dr\]

where

f(r,\phi,\theta) = exp-[\frac{(r^2sin^2\phi cos^2\theta)}{2\sigma_x^2}+\frac{(r^2sin^2\theta sin^2\phi)}{2\sigma_y^2}+\frac{r^2cos^2\phi}{2\sigma_z^2}]rdrd\theta

Cuas.createProjectile — Method

createProjectile(mass,Calibre;optional arguments)

Creates an object projectile with mass (kg), calibre (m). optional arguments are: position, velocity, time of flight, moment of inertia, distance to the center of gravity.

Cuas.createTourelle — Method

createTourelle(θ,ξ;optional arguments)

Creates an object turrets with elevation angle = θ and azimuth angle = ξ (in degrees). optional arguments are angles (in degrees): Φ, Ψ, χ and rwY.

Cuas.pHit — Method

pHit(target,error)

Returns the probability of hitting the target defined by'target' if the estimated error is provided by 'error'.

\[\int_{-a}^{a}\int_{-b}^{b}\frac{1}{2\pi\sigma_x\sigma_y}\text{exp}[\frac{(x-\mu_x)^2}{2\sigma_x^2}\frac{(y-\mu_y)^2}{2\sigma_y^2}]dxdy\]

Type documentation

Cuas.CEP

Cuas.CEP

Cuas.CEP

Cuas.MSHP

Cuas.SEP

Cuas.SSHP

Cuas.SSHP

Cuas.SSHP

Cuas.createProjectile

Cuas.createTourelle

Cuas.pHit