Cuas.jl Documentation

The anti aerial package.

Introduction

Cuas is a anti aerial simulation package designed to compute efficiency of a weapon system against an aerial target. It offers the following scenarios: stationary vs stationary, stationary vs moving, moving vs stationary and moving vs moving. For one on one, one on many, many on one and many on many configurations.

Package Features

Hit probability

Function Documentation

Cuas.pHit — Function

pHit(target,error)

Returns the probability of hitting the target defined by'target' if the estimated error is provided by 'error'.

\[\int_{-a}^{a}\int_{-b}^{b}\frac{1}{2\pi\sigma_x\sigma_y}\text{exp}[\frac{(x-\mu_x)^2}{2\sigma_x^2}\frac{(y-\mu_y)^2}{2\sigma_y^2}]dxdy\]

Cuas.createTourelle — Function

createTourelle(θ,ξ;optional arguments)

Creates an object turrets with elevation angle = θ and azimuth angle = ξ (in degrees). optional arguments are angles (in degrees): Φ, Ψ, χ and rwY.

Cuas.Turret — Type

Turret

The turret is defined by its azimuth angle

Cuas.createProjectile — Function

createProjectile(mass,Calibre;optional arguments)

Creates an object projectile with mass (kg), calibre (m). optional arguments are: position, velocity, time of flight, moment of inertia, distance to the center of gravity.